requirements.txt
Flask==3.0.0
gunicorn==21.2.0
mysql-connector-python==8.2.0
google-cloud-pubsub==2.18.4
